requirements.txt
absl-py==1.4.0
astunparse==1.6.3
boto3==1.35.58
botocore==1.35.58
cachetools==5.3.3
certifi>=2023.7.22
charset-normalizer>=3.3.2
click>=8.1.7
contourpy==1.1.1
cycler==0.12.1
filelock>=3.12.4
filetype==1.2.0
fire==0.5.0
flatbuffers==23.5.26
fonttools>=4.43.1
fsspec==2023.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.17.3
idna>=3.4
importlib-metadata==6.7.0
importlib-resources==6.1.1
iniconfig==2.0.0
Jinja2>=3.1.2
jmespath==1.0.1
joblib>=1.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown>=3.4.4
MarkupSafe>=2.1.3
matplotlib==3.7.5
mkl-service>=2.4.0
mpmath>=1.3.0
networkx>=3.1

numpy==1.24.4
oauthlib==3.2.2

opencv-python==4.9.0.80

opt_einsum==3.4.0
pafy==0.5.5
pandas==2.0.3
patsy==1.0.0
pillow>=10.0.0
pluggy==1.5.0
protobuf==3.20.3
psutil>=5.9.0

pytest==7.4.4

python-dateutil>=2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML>=6.0

pyzmq==25.1.2

requests>=2.31.0
requests-oauthlib>=1.3.1
requests-toolbelt>=1.0.0

roboflow==1.1.49
rsa==4.9
s3transfer==0.10.3
sahi==0.11.18

scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.1

six>=1.16.0
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tomli==2.0.1

torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2

tqdm>=4.66.1
typing_extensions>=4.7.1
tzdata==2023.3

ultralytics==8.0.221
ultralytics-thop==2.0.11
urllib3==1.26.18

wcwidth>=0.2.6
Werkzeug>=2.3.7
wrapt==1.15.0

yolov5==7.0.14
zipp>=3.15.0